requirements.txt
aiohttp==3.9.5
Flask==3.0.3
Flask_Login==0.6.3
flask_sqlalchemy==3.1.1
mysql_connector_repackaged==0.3.1
numpy==1.23.5
opencv_contrib_python==4.9.0.80
opencv_python==4.7.0.72
SQLAlchemy==2.0.30
Werkzeug==3.0.3
